Extract functionality to share between Address implementations
https://phabricator.wikimedia.org/T260023
